ai/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
ultralytics>=8.3.0
torch==2.11.0
torchvision==0.26.0
opencv-python-headless>=4.10.0
numpy>=2.2.0
pandas>=2.2.0
openpyxl>=3.1.0
requests>=2.32.0
python-dotenv>=1.0.0
cloudinary>=1.40.0
